requirements.txt
streamlit
python-docx
python-pptx
pypdf
pandas
